requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pywebview>=5.3
pydantic>=2.10
pyyaml>=6.0
ruamel.yaml>=0.18
requests>=2.32
mutagen>=1.47
pydub>=0.25
aiosqlite>=0.21
# Install separately with prebuilt wheel (build from source often fails on Windows):
#   pip install llama-cpp-python --prefer-binary --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
tqdm>=4.66
